Move audience field related methods
We now have OgGroupAudienceHelper class for this functionality
