+++ tests/src/Kernel/Entity/GroupAudienceTest.php
@@ -67,7 +67,7 @@ class GroupAudienceTest extends KernelTestBase {
     $bundle = $this->bundles[2];
 
     // Test no values returned for a non-group content.
-    $this->assertEmpty(Og::getAllGroupAudienceFields('entity_test', $bundle));
+    $this->assertEmpty(OgGroupAudienceHelper::getAllGroupAudienceFields('entity_test', $bundle));
 
     // Set bundles as group content.
     $field_name1 = Unicode::strtolower($this->randomMachineName());
@@ -76,11 +76,11 @@ class GroupAudienceTest extends KernelTestBase {
     Og::CreateField(OgGroupAudienceHelper::DEFAULT_FIELD, 'entity_test', $bundle, ['field_name' => $field_name1]);
     Og::CreateField(OgGroupAudienceHelper::DEFAULT_FIELD, 'entity_test', $bundle, ['field_name' => $field_name2]);
 
-    $field_names = Og::getAllGroupAudienceFields('entity_test', $bundle);
+    $field_names = OgGroupAudienceHelper::getAllGroupAudienceFields('entity_test', $bundle);
     $this->assertEquals(array($field_name1, $field_name2), array_keys($field_names));
 
     // Test Og::isGroupContent method, which is just a wrapper around
-    // Og::getAllGroupAudienceFields.
+    // OgGroupAudienceHelper::getAllGroupAudienceFields.
     $this->assertTrue(Og::isGroupContent('entity_test', $bundle));
 
     $bundle = $this->bundles[3];
@@ -112,7 +112,7 @@ class GroupAudienceTest extends KernelTestBase {
     // Add a default field, which will use the "entity_test" as target type.
     Og::CreateField(OgGroupAudienceHelper::DEFAULT_FIELD, 'entity_test', $bundle, ['field_name' => $field_name2]);
 
-    $field_names = Og::getAllGroupAudienceFields('entity_test', $bundle, 'entity_test');
+    $field_names = OgGroupAudienceHelper::getAllGroupAudienceFields('entity_test', $bundle, 'entity_test');
     $this->assertEquals(array($field_name2), array_keys($field_names));
 
   }
@@ -151,7 +151,7 @@ class GroupAudienceTest extends KernelTestBase {
     $overrides['field_config']['settings']['handler_settings']['target_bundles'] = [$group_bundle2 => $group_bundle2];
     Og::CreateField(OgGroupAudienceHelper::DEFAULT_FIELD, 'entity_test', $bundle, $overrides);
 
-    $field_names = Og::getAllGroupAudienceFields('entity_test', $bundle, 'entity_test', $group_bundle1);
+    $field_names = OgGroupAudienceHelper::getAllGroupAudienceFields('entity_test', $bundle, 'entity_test', $group_bundle1);
     $this->assertEquals(array($field_name1), array_keys($field_names));
   }
